+++ packages/lib-js-client/src/index.js
@@ -2,7 +2,7 @@ import querystring from 'querystring'
 import fetch from 'axios'
 import FormData from 'form-data'
 
-function SyncanoClient(instanceName = required('instanceName'), options = {}) {
+function SyncanoClient (instanceName = required('instanceName'), options = {}) {
   const host = options.host || 'syncano.space'
   const DEFAULT_HEADERS = {'Content-Type': 'application/json'}
 
@@ -14,26 +14,26 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
   client.headers = headers => Object.assign({}, DEFAULT_HEADERS, headers)
   client.post = client
 
-  client.login = function(username, password) {
+  client.login = function (username, password) {
     const login = client.loginMethod
       ? client.loginMethod
       : (username, password) => {
-          const url = `https://api.syncano.io/v2/instances/${
+        const url = `https://api.syncano.io/v2/instances/${
             client.instanceName
           }/users/auth/`
-          const data = JSON.stringify({username, password})
+        const data = JSON.stringify({username, password})
 
-          return fetch({url, data}).then(user => {
-            client.setToken(user.token)
+        return fetch({url, data}).then(user => {
+          client.setToken(user.token)
 
-            return user
-          })
-        }
+          return user
+        })
+      }
 
     return login(username, password)
   }
 
-  client.url = function(endpoint, data) {
+  client.url = function (endpoint, data) {
     let url = `${this.baseUrl}${endpoint}/?`
 
     if (data) {
@@ -47,11 +47,11 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     return url
   }
 
-  client.logout = function() {
+  client.logout = function () {
     this.token = undefined
   }
 
-  client.setToken = function(token) {
+  client.setToken = function (token) {
     this.token = token
 
     if (typeof client.setTokenCallback === 'function') {
@@ -59,7 +59,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     }
   }
 
-  client.get = function(
+  client.get = function (
     endpoint = required('endpoint'),
     data = {},
     options = {}
@@ -67,7 +67,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     return this.post(endpoint, {...data, _method: 'GET'}, options)
   }
 
-  client.delete = function(
+  client.delete = function (
     endpoint = required('endpoint'),
     data = {},
     options = {}
@@ -75,7 +75,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     return this.post(endpoint, {...data, _method: 'DELETE'}, options)
   }
 
-  client.put = function(
+  client.put = function (
     endpoint = required('endpoint'),
     data = {},
     options = {}
@@ -83,7 +83,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     return this.post(endpoint, {...data, _method: 'PUT'}, options)
   }
 
-  client.patch = function(
+  client.patch = function (
     endpoint = required('endpoint'),
     data = {},
     options = {}
@@ -92,7 +92,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
   }
 
   // Used by the client.subscribe method to start polling from the correct id
-  client.setLastId = function(endpoint, data) {
+  client.setLastId = function (endpoint, data) {
     const url = this.url(`${endpoint}/history`, data)
     // eslint-disable-next-line camelcase
     if (data.last_id) {
@@ -105,7 +105,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     })
   }
 
-  client.subscribe = function(endpoint = required('endpoint'), data, callback) {
+  client.subscribe = function (endpoint = required('endpoint'), data, callback) {
     let abort = false
     const hasData = typeof data === 'object' && data !== null
 
@@ -124,7 +124,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     let url = client.url(endpoint, data)
     const cb = hasData ? callback : data
 
-    function loop() {
+    function loop () {
       if (abort) {
         return
       }
@@ -161,7 +161,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     }
   }
 
-  client.subscribe.once = function(
+  client.subscribe.once = function (
     endpoint = required('endpoint'),
     data = {},
     callback
@@ -180,7 +180,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
 
   return client
 
-  function client(endpoint = required('endpoint'), data = {}, options = {}) {
+  function client (endpoint = required('endpoint'), data = {}, options = {}) {
     const method = data._method || 'POST'
     const headers = client.headers(options.headers)
     let url = client.url(endpoint)
@@ -195,7 +195,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
     }
 
     const transformRequest = [
-      function(data) {
+      function (data) {
         if (data instanceof FormData) {
           return data
         }
@@ -215,7 +215,7 @@ function SyncanoClient(instanceName = required('instanceName'), options = {}) {
   }
 }
 
-function required(param) {
+function required (param) {
   throw new Error(`${param} parameter is required by SyncanoClient`)
 }