+++ packages/test-tools/src/index.js
@@ -1,4 +1,5 @@
 /* eslint import/no-extraneous-dependencies: "warn" */
+import faker from 'faker'
 import mkdirp from 'mkdirp'
 import fs from 'fs-extra'
 import path from 'path'
@@ -13,10 +14,11 @@ import homeDir from 'home-dir'
 const returnTestGlobals = () => {
   return {
     email: process.env.E2E_CLI_EMAIL,
+    apiHost: process.env.E2E_SYNCANO_HOST,
     password: process.env.E2E_CLI_PASSWORD,
     accountKey: process.env.E2E_CLI_ACCOUNT_KEY,
     syncanoYmlPath: `${homeDir()}/syncano-test.yml`,
-    instance: process.env.E2E_CLI_INSTANCE_NAME || 'wandering-pine-7032'
+    instance: process.env.E2E_CLI_INSTANCE_NAME
   }
 }
 
@@ -30,7 +32,6 @@ const splitTestBaseEmail = (tempEmail) => {
 
 const createTempEmail = (tempEmail, tempPass) => {
   const { emailName, emailDomain } = splitTestBaseEmail(tempEmail)
-
   return `${emailName}+${tempPass}@${emailDomain}`
 }
 
@@ -61,13 +62,10 @@ if (process.env.SYNCANO_E2E_DEBUG) {
 }
 const nixt = origNixt
 
-process.env.SYNCANO_ACCOUNT_FILE = 'syncano-test'
-
 // Variables used in tests
-const { accountKey, syncanoYmlPath, instance } = returnTestGlobals()
+process.env.SYNCANO_ACCOUNT_FILE = 'syncano-test'
+const { syncanoYmlPath, instance } = returnTestGlobals()
 
-// const connection = Syncano({ baseUrl: `https://${process.env.SYNCANO_HOST}`, accountKey })
-const connection = new Syncano({accountKey, meta: { 'api_host': process.env.SYNCANO_HOST }})
 const testsLocation = `${process.cwd()}/e2e-tests`
 const cliLocation = path.join(process.cwd(), '/node_modules/@syncano/cli/lib/cli.js')
 const randomKey = getRandomString()
@@ -88,16 +86,16 @@ const shutdownLocation = (location) => {
   fs.removeSync(syncanoYmlPath)
 }
 
-// Helper functions used in tests
-const createInstance = (instanceName) => connection.instance
-    .create({ name: instanceName || uniqueInstance() })
-    .catch((error) => process.stderr.write(JSON.stringify(error.message, null, '')))
+const createConnection = () => {
+  const {accountKey, apiHost} = returnTestGlobals()
+  return new Syncano({accountKey, meta: {api_host: apiHost}})
+}
 
-const deleteInstance = (instanceName) => connection.instance
-    .delete(instanceName)
-    .catch((error) => process.stderr.write(
-      JSON.stringify(`deleteInstance: ${error.message}`, null, '')
-    ))
+// Helper functions used in tests
+const createInstance = (instanceName) =>
+  createConnection().instance.create({ name: instanceName || uniqueInstance() })
+
+const deleteInstance = (instanceName) => createConnection().instance.delete(instanceName)
 
 const deleteEachInstance = (instances) => {
   const list = []
@@ -106,20 +104,106 @@ const deleteEachInstance = (instances) => {
   return Promise.all(list)
 }
 
-const cleanUpAccount = () => connection.instance
+const cleanUpAccount = () =>
+  createConnection().instance
     .list()
     .then(res => {
       const instances = _.pull(_.map(res, 'name'), instance)
       return deleteEachInstance(instances)
     })
-    .catch((error) => process.stderr.write(JSON.stringify(error.message)))
 
 const createProject = (instanceName, projectTestTemplate) => {
   fs.copySync(projectTestTemplate, path.join(testsLocation, instanceName))
   return createInstance(instanceName)
 }
 
+const generateResponse = (response) => {
+  return Object.assign({
+    code: 200,
+    data: {},
+    mimetype: 'application/json'
+  }, response)
+}
+
+const generateContext = () => {
+  return {
+    meta:
+    {
+      socket: 'norwegian-postcode',
+      api_host: 'api.syncano.io',
+      token: 'token',
+      instanc: 'syncnao-instance',
+      debug: process.env.DEBUG || false,
+      executor: 'norwegian-postcode/search',
+      executed_by: 'socket_endpoint',
+      request: {
+        REQUEST_METHOD: 'POST',
+        PATH_INFO: '/v2/instances/withered-voice-2245/endpoints/sockets/norwegian-postcode/search/',
+        HTTP_USER_AGENT: faker.internet.userAgent(),
+        HTTP_CONNECTION: 'close',
+        REMOTE_ADDR: faker.internet.ip(),
+        HTTP_HOST: 'api.syncano.io',
+        HTTP_UPGRADE_INSECURE_REQUESTS: '1',
+        HTTP_ACCEPT: 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8',
+        HTTP_ACCEPT_LANGUAGE: 'en,en-US;q=0.8,pl;q=0.6',
+        HTTP_ACCEPT_ENCODING: 'gzip, deflate, br'
+      },
+      metadata: {
+        description: 'Search for data based on given postcode',
+        inputs: {
+          type: 'object',
+          properties: {
+            postcode: {
+              type: 'integer',
+              description: 'Post code',
+              example: 113
+            }
+          }
+        },
+        outputs: {
+          success: {
+            description: 'Successful query',
+            properties: {
+              city: {
+                typeOf: 'string'
+              },
+              municipality: {
+                type: 'string'
+              },
+              county: {
+                'type': 'string'
+              },
+              category: {
+                type: 'string'
+              }
+            }
+          },
+          notFound: {
+            description: 'Post code not found',
+            exit_code: 404,
+            properties: {
+              message: {
+                type: 'string'
+              }
+            }
+          },
+          fail: {
+            description: 'Zip generation failed',
+            exit_code: 400,
+            parameters: {
+              message: {
+                type: 'string'
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
 export {
+  createConnection,
   testsLocation,
   returnTestGlobals,
   getRandomString,
@@ -137,5 +221,7 @@ export {
   createInstance,
   deleteInstance,
   uniqueInstance,
-  cleanUpAccount
+  cleanUpAccount,
+  generateContext,
+  generateResponse
 }